Introduce ripport for importing
Jason created antitariff which this script is heavily based on. ripport
stands for ripgrep-import which will use ripgrep to find and count
usages of a text object and use the most common statement for an import
statement.

If ripgrep is not installed, ripport will fallback to vanilla grep but
of course the speed that grep works at is considerably worse if
navigating the gradle caches. If we can find the imports in the local
repository, then grep is usually fast enough.