Make ctrl+p use Telescope git_files
git_files will search the entirety of the repo without needing to set a
search root, as opposed to find_files which works from the current
directory. Instead of trying to configure a rooter, I thought I'd try
using git_files for a bit and see how I like it.